Make the quick sort work in the linux list style
In original code, it needs to use array end to check whether it contain
one element, but it is actually useless, I change it to save the length
of the list. Besides, I go throught the list with list_for_each_entry_safe
instead of the while loop. And I free all the elements which are created
by malloc.